Minor clean up.

+++ app/src/androidTest/java/com/example/studious/NotificationHelperTest.java
@@ -19,7 +19,7 @@ public class NotificationHelperTest {
     public void sendCalendarNotification() {
         Context context = InstrumentationRegistry.getInstrumentation().getTargetContext();
         NotificationHelper notificationHelper = new NotificationHelper(context);
-        Items item = new Items(17, "calendar title", Items.TYPE_CALENDAR, "here","12:30", "13:00","04/15/2022","04/16/2022");
+        Items item = new Items(17, "calendar title", Items.TYPE_CALENDAR, "here", "12:30", "13:00", "04/15/2022", "04/16/2022");
         notificationHelper.sendCalendarNotification(item);
     }
 
@@ -27,7 +27,7 @@ public class NotificationHelperTest {
     public void sendTodoNotification() {
         Context context = InstrumentationRegistry.getInstrumentation().getTargetContext();
         NotificationHelper notificationHelper = new NotificationHelper(context);
-        Items item = new Items(18, "todo title", Items.TYPE_TODO, "here","12:30", "13:00","04/15/2022","04/16/2022");
+        Items item = new Items(18, "todo title", Items.TYPE_TODO, "here", "12:30", "13:00", "04/15/2022", "04/16/2022");
         notificationHelper.sendTodoNotification(item);
     }
 
@@ -35,7 +35,7 @@ public class NotificationHelperTest {
     public void sendReminderNotification() {
         Context context = InstrumentationRegistry.getInstrumentation().getTargetContext();
         NotificationHelper notificationHelper = new NotificationHelper(context);
-        Items item = new Items(19, "reminder title", Items.TYPE_REMINDER, "here","12:30", "13:00","04/15/2022","04/16/2022");
+        Items item = new Items(19, "reminder title", Items.TYPE_REMINDER, "here", "12:30", "13:00", "04/15/2022", "04/16/2022");
         notificationHelper.sendReminderNotification(item);
     }
 
@@ -43,7 +43,7 @@ public class NotificationHelperTest {
     public void sendAlarmNotification() {
         Context context = InstrumentationRegistry.getInstrumentation().getTargetContext();
         NotificationHelper notificationHelper = new NotificationHelper(context);
-        Items item = new Items(19, "alarm title", Items.TYPE_ALARM, "here","12:30", "13:00","04/15/2022","04/16/2022");
+        Items item = new Items(19, "alarm title", Items.TYPE_ALARM, "here", "12:30", "13:00", "04/15/2022", "04/16/2022");
         notificationHelper.sendAlarmNotification(item);
     }
 
@@ -56,8 +56,8 @@ public class NotificationHelperTest {
 
         Context context = InstrumentationRegistry.getInstrumentation().getTargetContext();
         DataBaseHelper databaseHelper = new DataBaseHelper(context);
-        Items newItem = new Items(-1, "fullStackCalendar", Items.TYPE_CALENDAR, "here","12:30", "13:00","04/15/2022","04/16/2022");
-        int row_id = (int)databaseHelper.addItem(newItem);
+        Items newItem = new Items(-1, "fullStackCalendar", Items.TYPE_CALENDAR, "here", "12:30", "13:00", "04/15/2022", "04/16/2022");
+        int row_id = databaseHelper.addItem(newItem);
 
         Calendar c = Calendar.getInstance();
         AlarmMangerHelper alarmMangerHelper = new AlarmMangerHelper(context);
@@ -81,8 +81,8 @@ public class NotificationHelperTest {
 
         Context context = InstrumentationRegistry.getInstrumentation().getTargetContext();
         DataBaseHelper databaseHelper = new DataBaseHelper(context);
-        Items newItem = new Items(-1, "fullStackReminder", Items.TYPE_REMINDER, "here","12:30", "13:00","04/15/2022","04/16/2022");
-        int row_id = (int)databaseHelper.addItem(newItem);
+        Items newItem = new Items(-1, "fullStackReminder", Items.TYPE_REMINDER, "here", "12:30", "13:00", "04/15/2022", "04/16/2022");
+        int row_id = databaseHelper.addItem(newItem);
 
         Calendar c = Calendar.getInstance();
         AlarmMangerHelper alarmMangerHelper = new AlarmMangerHelper(context);
@@ -106,8 +106,8 @@ public class NotificationHelperTest {
 
         Context context = InstrumentationRegistry.getInstrumentation().getTargetContext();
         DataBaseHelper databaseHelper = new DataBaseHelper(context);
-        Items newItem = new Items(-1, "fullStackAlarm", Items.TYPE_ALARM, "here","12:30", "13:00","04/15/2022","04/16/2022");
-        int row_id = (int)databaseHelper.addItem(newItem);
+        Items newItem = new Items(-1, "fullStackAlarm", Items.TYPE_ALARM, "here", "12:30", "13:00", "04/15/2022", "04/16/2022");
+        int row_id = databaseHelper.addItem(newItem);
 
         Calendar c = Calendar.getInstance();
         AlarmMangerHelper alarmMangerHelper = new AlarmMangerHelper(context);
@@ -131,8 +131,8 @@ public class NotificationHelperTest {
 
         Context context = InstrumentationRegistry.getInstrumentation().getTargetContext();
         DataBaseHelper databaseHelper = new DataBaseHelper(context);
-        Items newItem = new Items(-1, "fullStackTodo", Items.TYPE_TODO, "here","12:30", "13:00","04/15/2022","04/16/2022");
-        int row_id = (int)databaseHelper.addItem(newItem);
+        Items newItem = new Items(-1, "fullStackTodo", Items.TYPE_TODO, "here", "12:30", "13:00", "04/15/2022", "04/16/2022");
+        int row_id = databaseHelper.addItem(newItem);
 
         Calendar c = Calendar.getInstance();
         AlarmMangerHelper alarmMangerHelper = new AlarmMangerHelper(context);
